cv-chat/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==3.7.1
async-timeout==5.0.1
attrs==25.3.0
bcrypt==4.3.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
dnspython==2.7.0
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.104.1
filelock==3.18.0
fonttools==4.58.2
frozenlist==1.7.0
fsspec==2025.5.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
motor==3.7.1
mpmath==1.3.0
multidict==6.4.4
networkx==3.4.2
numpy==1.24.3
opencv-python==4.8.1.78
packaging==25.0
pandas==2.3.0
passlib==1.7.4
Pillow==10.0.1
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.4.2
pydantic_core==2.10.1
pymongo==4.13.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rsa==4.9.1
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
sympy==1.14.0
thop==0.1.1.post2209072238
torch==2.1.0
torchvision==0.16.0
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
ultralytics==8.3.153
ultralytics-thop==2.0.14
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.24.0
watchfiles==1.0.5
websockets==15.0.1
yarl==1.20.1
